Move StarsReviews.jsx into common folder and rename StarsContainer so it can be used by all (moved styles for stars our of ProductInfoStyles.jsx and into StarsContainerStyles.jsx) - StarsContainer takes a boolean prop which determines whether it should also generate the reviews list as well(which is what I need for the product overview component) - could be de-coupled and made more modular if time
